rename message_batch_helpers
we're going to add another method in here so
think it warrants its own batch-specific file
