+++ manage_breast_screening/notifications/management/commands/retry_failed_message_batch.py
@@ -5,7 +5,7 @@ from logging import getLogger
 
 from django.core.management.base import BaseCommand, CommandError
 
-from manage_breast_screening.notifications.management.commands.command_helpers import (
+from manage_breast_screening.notifications.management.commands.helpers.message_batch_helpers import (
     MessageBatchHelpers,
 )
 from manage_breast_screening.notifications.models import (
